+++ pext/__main__.py
@@ -906,7 +906,15 @@ class ViewModel():
         current_index = QQmlProperty.read(self.result_list_model, "currentIndex")
 
         if self.result_list_model_command_mode or len(self.filtered_entry_list) == 0:
+            selected_command = self.filtered_command_list[current_index - len(self.filtered_entry_list)]
+            selected_command_split = selected_command.split(" ", 1)
             command_typed = QQmlProperty.read(self.search_input_model, "text")
+            command_typed_split = command_typed.split(" ", 1)
+
+            try:
+                command_typed = selected_command_split[0] + " " + command_typed_split[1]
+            except IndexError:
+                command_typed = selected_command_split[0]
 
             self.selection.append(
                 {'type': SelectionType.command, 'value': command_typed})